Refactor GetGlobalDdevDir() into util to avoid import cycles
