+++ pkg/ddevapp/config.go
@@ -16,7 +16,6 @@ import (
 	"github.com/drud/ddev/pkg/appports"
 	"github.com/drud/ddev/pkg/util"
 	"github.com/drud/ddev/pkg/version"
-	homedir "github.com/mitchellh/go-homedir"
 	"github.com/pkg/errors"
 	yaml "gopkg.in/yaml.v2"
 )
@@ -158,12 +157,7 @@ func (c *Config) Read() error {
 		c.DBAImage = version.DBAImg + ":" + version.DBATag
 	}
 
-	dirPath, err := homedir.Dir()
-	if err != nil {
-		return err
-	}
-
-	dirPath = filepath.Join(dirPath, ".ddev", c.Name)
+	dirPath := filepath.Join(util.GetGlobalDdevDir(), c.Name)
 	c.DataDir = filepath.Join(dirPath, "mysql")
 	c.ImportDir = filepath.Join(dirPath, "import-db")
